Add verify-publishable bin
