+++ semantic-release/base.js
@@ -10,38 +10,11 @@
  * }} SemanticReleaseConfig
  */
 
-import { readFileSync } from "node:fs";
-import { createRequire } from "node:module";
-import { dirname, join } from "node:path";
+import { fileURLToPath } from "node:url";
 
-const requireFromHere = createRequire(import.meta.url);
-
-/**
- * Resolves a package's bin script from this package's dependency tree so
- * release checks run the installed copy and never trigger an npx download.
- *
- * @param {string} pkg
- * @param {string} [binName]
- * @returns {string}
- */
-function resolveBin(pkg, binName = pkg) {
-  for (const base of requireFromHere.resolve.paths(pkg) ?? []) {
-    const pkgJsonPath = join(base, pkg, "package.json");
-    try {
-      const { bin } = JSON.parse(readFileSync(pkgJsonPath, "utf8"));
-      return join(
-        dirname(pkgJsonPath),
-        typeof bin === "string" ? bin : bin[binName],
-      );
-    } catch {
-      // not in this node_modules dir, try the next candidate
-    }
-  }
-  throw new Error(`Cannot resolve bin for "${pkg}"`);
-}
-
-const publintBin = resolveBin("publint");
-const attwBin = resolveBin("@arethetypeswrong/cli", "attw");
+const verifyPublishableScript = fileURLToPath(
+  new URL("verify-publishable.js", import.meta.url),
+);
 
 /** @type {SemanticReleaseConfig} */
 export const config = {
@@ -62,7 +35,7 @@ export const config = {
       {
         verifyConditionsCmd:
           "node -e \"const fs=require('node:fs'); if (process.env.SEMANTIC_RELEASE_ALLOW_CHANGELOG === 'true') process.exit(0); if (fs.existsSync('CHANGELOG.md')) { console.error('Release aborted: CHANGELOG.md is present. Remove it or set SEMANTIC_RELEASE_ALLOW_CHANGELOG=true to bypass this check.'); process.exit(1); }\"" +
-          ` && node "${publintBin}" && node "${attwBin}" --pack . --ignore-rules no-resolution cjs-resolves-to-esm`,
+          ` && node "${verifyPublishableScript}"`,
       },
     ],
     "@semantic-release/commit-analyzer",
